[Feat] #18 - 뷰 구조화 & ScanView 수정중
