+++ PennyPack/Source/View/MainView.swift
@@ -1,95 +1,138 @@
-//
-//  MainView.swift
-//  PennyPack
-//
-//  Created by siye on 11/1/24.
-//
-
-import Foundation
 import SwiftUI
 
-func pricing(_ text: String) -> String {
-    let pattern = #"^\d{1,2}\.\d{1,2}$"#
-    let regex = try! NSRegularExpression(pattern: pattern, options: [])
-        let range = NSRange(location: 0, length: text.utf16.count)
-        let cleanedText = regex.stringByReplacingMatches(in: text, options: [], range: range, withTemplate: "")
-        return cleanedText.trimmingCharacters(in: .whitespacesAndNewlines)
-}
-    let prices = ["4.33", "4€ .33", "14.33"]
-
-// 데이터 정제 및 토큰화
-func tokenizeAndLabel(_ text: String) -> [(String, String)] {
-    let priced = pricing(text)
-    let tokens = priced.components(separatedBy: .whitespacesAndNewlines).filter { !$0.isEmpty }
-    
-    // 토큰에 라벨 붙이기
-    return tokens.map { token in
-        if let _ = Double(token) {
-            return (token, "NUMBER")
-        } else if token.lowercased() == "€" {
-            return (token, "CURRENCY")
-        } else {
-            return (token, "TEXT")
-        }
-    }
-}
-
-// JSON 생성
-func createJSONForCreateML(_ labeledTokens: [(String, String)]) -> String {
-    let jsonObject: [String: Any] = [
-        "tokens": labeledTokens.map { $0.0 },
-        "labels": labeledTokens.map { $0.1 }
-    ]
-    
-    let jsonData = try! JSONSerialization.data(withJSONObject: jsonObject, options: .prettyPrinted)
-    return String(data: jsonData, encoding: .utf8)!
-}
-
-// 4. 전체 프로세스
-func processTextForCreateML(_ text: String) -> String {
-    let labeledTokens = tokenizeAndLabel(text)
-    return createJSONForCreateML(labeledTokens)
-}
-
-// 사용 예시
-let inputText = "The price is 14.99 € for this item"
-let jsonOutput = processTextForCreateML(inputText)
-
-func printjson() {
-    print(jsonOutput)
-}
-
 struct MainView: View {
-    @ObservedObject var currencyService = CurrencyService()
-    @State private var selectedCurrencyType: CurrencyType = .usd  // 기본 선택 통화 설정
-
+//    @StateObject private var shoppingViewModel = ShoppingViewModel()
+    
+    @ObservedObject var shoppingViewModel: ShoppingViewModel
+    @ObservedObject var linkListStore: LinkListStore = LinkListStore()
+   
+    @State private var listText = ""
     var body: some View {
-        VStack {
-            Picker("Select Currency", selection: $selectedCurrencyType) {
-                ForEach(CurrencyType.allCases) { currencyType in
-                    Text(currencyType.rawValue).tag(currencyType)
+        NavigationStack{
+            ZStack{
+                Color.rBackgroundGray
+                    .ignoresSafeArea()
+                VStack{
+                    Color.rBlack
+                        .ignoresSafeArea()
+                        .frame(height: 339)
+                    Spacer()
+                }
+                
+                VStack(spacing: 24){
+                    HStack{
+                        Text("PennyPack")
+                            .font(.RMainTitle)
+                            .foregroundColor(.white)
+                        Spacer()
+                    }.padding(.horizontal)
+                    
+                    HStack{
+                        Image("France")
+                            .resizable()
+                            .scaledToFit()
+                            .frame(width: 35)
+                        Text("프랑스")
+                        Spacer()
+                        Text("€ 1 = ₩ 1499.62")
+                        Text("(EUR/KRW)")
+                    }
+                    .padding(.vertical, 12)
+                    .padding(.horizontal)
+                    .background(
+                        Color.white
+                            .cornerRadius(12)
+                    )
+                    .overlay(
+                        RoundedRectangle(cornerRadius: 12)
+                            .stroke(Color.rStrokeGray, lineWidth: 1)
+                        
+                    )
+                    .padding(.horizontal)
+                    
+                    VStack(spacing: 0){
+                        HStack{
+                            Text("오늘의 장보기 리스트")
+                            Spacer()
+                        }
+                        .padding()
+                        .background(
+                            Rectangle()
+                                .foregroundColor(.white)
+                                .clipShape(RoundedCorner(radius: 12, corners: [.topLeft, .topRight]))
+                        )
+                        .padding(.horizontal)
+                        ZStack{
+                            Rectangle()
+                                .foregroundColor(.rLightGray)
+                                .clipShape(RoundedCorner(radius: 12, corners: [.bottomLeft, .bottomRight]))
+                            
+                            VStack(spacing: 0){
+                                List{
+                                    ForEach($linkListStore.linkLists, id: \.id){ $item in
+                                        TextField("마트에서 살 물건을 이곳에 적어주세요.", text: $item.title)
+                                        
+                                    }
+                                    .listRowBackground(
+                                        Rectangle()
+                                            .foregroundColor(.white)
+                                            .cornerRadius(12)
+                                    )
+                                    Button {
+                                        linkListStore.addLinkList(title: listText)
+                                        
+                                    } label: {
+                                        HStack{
+                                            Spacer()
+                                            Image(systemName: "plus.circle.fill")
+                                                .font(.system(size: 18))
+                                                .foregroundStyle(.gray, .rBackgroundGray)
+                                            Spacer()
+                                        }
+                                    }
+                                    .listRowBackground(
+                                        Rectangle()
+                                            .foregroundColor(.white)
+                                            .cornerRadius(12)
+                                    )
+                                }.listRowSpacing(8)
+                                    .listStyle(PlainListStyle())
+                                    .padding(.horizontal)
+                                    .background(.rLightGray)
+                                
+                                
+                            }.padding(.vertical)
+                        }.padding(.horizontal)
+                        NavigationLink(
+                            destination: CartView(shoppingViewModel: shoppingViewModel),
+                            label: {
+                                HStack{
+                                    Image(systemName: "cart")
+                                        .font(.system(size: 24))
+                                    Text("장보기 시작")
+                                }.foregroundColor(.white)
+                                    .padding(.vertical,13)
+                                    .padding(.horizontal,111)
+                                .background(
+                                    Rectangle()
+                                        .foregroundColor(.rMainBlue)
+                                        .cornerRadius(24)
+                                )
+                            })
+                        
+                        .padding(.top,24)
+                        .padding(.bottom,30)
+                    }
                 }
             }
-            .pickerStyle(.menu)
-            .onChange(of: selectedCurrencyType) { newCurrencyType in
-                currencyService.getRate(for: newCurrencyType.cur_unit)
-            }
-
-            if let rate = currencyService.selectedRate {
-                Text("Exchange Rate for \(selectedCurrencyType.rawValue): \(String(format: "%.0f", rate))")
-                    .padding()
-            } else {
-                Text("Select a currency to see the rate")
-                    .padding()
-            }
-        }
-        .onAppear {
-            currencyService.fetchCurrencies()
-            currencyService.getRate(for: selectedCurrencyType.cur_unit)
         }
     }
 }
 
+
+
+
 #Preview {
-    MainView()
+    MainView(shoppingViewModel: ShoppingViewModel())
 }
+
