+++ PennyPack/ViewModel/OtherFuncs.swift
@@ -15,4 +15,3 @@ struct RoundedCorner: Shape {
         return Path(path.cgPath)
     }
 }
-
